moved Upload button to separate component
